requirements.txt
numpy
Pillow
pkg-resources
protobuf
six
tensorboardX
torch
torchvision
torchvision
tqdm
opencv-python
imageio
tensorboard
packaging